agentcore_deployment/requirements.txt
# AgentCore and Bedrock
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3>=1.40.0

# Multi-Agent Frameworks
strands-agents
strands-agents-tools
langgraph>=0.6.6
crewai>=0.177.0
crewai-tools>=0.0.1

# AWS and LangChain Integration
langchain-aws>=0.2.31
langchain-core>=0.3.75
langchain-community

# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0,<2.0.0

# Data Processing and Validation
pydantic>=2.11.0
pandas>=2.3.0

# Utilities
requests>=2.32.0
python-dotenv>=1.1.0
tqdm>=4.67.0
